# [pkg/ottl] Use generics to avoid context cast in getters and funcs