--- a/pkg/ottl/compare.go
+++ b/pkg/ottl/compare.go
@@ -28,7 +28,7 @@ import (
 // invalidComparison returns false for everything except NE (where it returns true to indicate that the
 // objects were definitely not equivalent).
 // It also gives us an opportunity to log something.
-func (p *Parser) invalidComparison(msg string, op CompareOp) bool {
+func (p *Parser[K]) invalidComparison(msg string, op CompareOp) bool {
 	p.telemetrySettings.Logger.Debug(msg, zap.Any("op", op))
 	return op == NE
 }
@@ -92,7 +92,7 @@ func compareBytes(a []byte, b []byte, op CompareOp) bool {
 	}
 }
 
-func (p *Parser) compareBool(a bool, b any, op CompareOp) bool {
+func (p *Parser[K]) compareBool(a bool, b any, op CompareOp) bool {
 	switch v := b.(type) {
 	case bool:
 		return compareBools(a, v, op)
@@ -101,7 +101,7 @@ func (p *Parser) compareBool(a bool, b any, op CompareOp) bool {
 	}
 }
 
-func (p *Parser) compareString(a string, b any, op CompareOp) bool {
+func (p *Parser[K]) compareString(a string, b any, op CompareOp) bool {
 	switch v := b.(type) {
 	case string:
 		return comparePrimitives(a, v, op)
@@ -110,7 +110,7 @@ func (p *Parser) compareString(a string, b any, op CompareOp) bool {
 	}
 }
 
-func (p *Parser) compareByte(a []byte, b any, op CompareOp) bool {
+func (p *Parser[K]) compareByte(a []byte, b any, op CompareOp) bool {
 	switch v := b.(type) {
 	case nil:
 		return op == NE
@@ -124,7 +124,7 @@ func (p *Parser) compareByte(a []byte, b any, op CompareOp) bool {
 	}
 }
 
-func (p *Parser) compareInt64(a int64, b any, op CompareOp) bool {
+func (p *Parser[K]) compareInt64(a int64, b any, op CompareOp) bool {
 	switch v := b.(type) {
 	case int64:
 		return comparePrimitives(a, v, op)
@@ -135,7 +135,7 @@ func (p *Parser) compareInt64(a int64, b any, op CompareOp) bool {
 	}
 }
 
-func (p *Parser) compareFloat64(a float64, b any, op CompareOp) bool {
+func (p *Parser[K]) compareFloat64(a float64, b any, op CompareOp) bool {
 	switch v := b.(type) {
 	case int64:
 		return comparePrimitives(a, float64(v), op)
@@ -148,7 +148,7 @@ func (p *Parser) compareFloat64(a float64, b any, op CompareOp) bool {
 
 // a and b are the return values from a Getter; we try to compare them
 // according to the given operator.
-func (p *Parser) compare(a any, b any, op CompareOp) bool {
+func (p *Parser[K]) compare(a any, b any, op CompareOp) bool {
 	// nils are equal to each other and never equal to anything else,
 	// so if they're both nil, report equality.
 	if a == nil && b == nil {