--- a/pkg/ottl/contexts/internal/ottlcommon/scope.go
+++ b/pkg/ottl/contexts/internal/ottlcommon/scope.go
@@ -22,33 +22,37 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl"
 )
 
-func ScopePathGetSetter(path []ottl.Field) (ottl.GetSetter, error) {
+type InstrumentationScopeContext interface {
+	GetInstrumentationScope() pcommon.InstrumentationScope
+}
+
+func ScopePathGetSetter[K InstrumentationScopeContext](path []ottl.Field) (ottl.GetSetter[K], error) {
 	if len(path) == 0 {
-		return accessInstrumentationScope(), nil
+		return accessInstrumentationScope[K](), nil
 	}
 
 	switch path[0].Name {
 	case "name":
-		return accessInstrumentationScopeName(), nil
+		return accessInstrumentationScopeName[K](), nil
 	case "version":
-		return accessInstrumentationScopeVersion(), nil
+		return accessInstrumentationScopeVersion[K](), nil
 	case "attributes":
 		mapKey := path[0].MapKey
 		if mapKey == nil {
-			return accessInstrumentationScopeAttributes(), nil
+			return accessInstrumentationScopeAttributes[K](), nil
 		}
-		return accessInstrumentationScopeAttributesKey(mapKey), nil
+		return accessInstrumentationScopeAttributesKey[K](mapKey), nil
 	}
 
 	return nil, fmt.Errorf("invalid scope path expression %v", path)
 }
 
-func accessInstrumentationScope() ottl.StandardGetSetter {
-	return ottl.StandardGetSetter{
-		Getter: func(ctx ottl.TransformContext) interface{} {
+func accessInstrumentationScope[K InstrumentationScopeContext]() ottl.StandardGetSetter[K] {
+	return ottl.StandardGetSetter[K]{
+		Getter: func(ctx K) interface{} {
 			return ctx.GetInstrumentationScope()
 		},
-		Setter: func(ctx ottl.TransformContext, val interface{}) {
+		Setter: func(ctx K, val interface{}) {
 			if newIl, ok := val.(pcommon.InstrumentationScope); ok {
 				newIl.CopyTo(ctx.GetInstrumentationScope())
 			}
@@ -56,12 +60,12 @@ func accessInstrumentationScope() ottl.StandardGetSetter {
 	}
 }
 
-func accessInstrumentationScopeAttributes() ottl.StandardGetSetter {
-	return ottl.StandardGetSetter{
-		Getter: func(ctx ottl.TransformContext) interface{} {
+func accessInstrumentationScopeAttributes[K InstrumentationScopeContext]() ottl.StandardGetSetter[K] {
+	return ottl.StandardGetSetter[K]{
+		Getter: func(ctx K) interface{} {
 			return ctx.GetInstrumentationScope().Attributes()
 		},
-		Setter: func(ctx ottl.TransformContext, val interface{}) {
+		Setter: func(ctx K, val interface{}) {
 			if attrs, ok := val.(pcommon.Map); ok {
 				attrs.CopyTo(ctx.GetInstrumentationScope().Attributes())
 			}
@@ -69,23 +73,23 @@ func accessInstrumentationScopeAttributes() ottl.StandardGetSetter {
 	}
 }
 
-func accessInstrumentationScopeAttributesKey(mapKey *string) ottl.StandardGetSetter {
-	return ottl.StandardGetSetter{
-		Getter: func(ctx ottl.TransformContext) interface{} {
+func accessInstrumentationScopeAttributesKey[K InstrumentationScopeContext](mapKey *string) ottl.StandardGetSetter[K] {
+	return ottl.StandardGetSetter[K]{
+		Getter: func(ctx K) interface{} {
 			return GetMapValue(ctx.GetInstrumentationScope().Attributes(), *mapKey)
 		},
-		Setter: func(ctx ottl.TransformContext, val interface{}) {
+		Setter: func(ctx K, val interface{}) {
 			SetMapValue(ctx.GetInstrumentationScope().Attributes(), *mapKey, val)
 		},
 	}
 }
 
-func accessInstrumentationScopeName() ottl.StandardGetSetter {
-	return ottl.StandardGetSetter{
-		Getter: func(ctx ottl.TransformContext) interface{} {
+func accessInstrumentationScopeName[K InstrumentationScopeContext]() ottl.StandardGetSetter[K] {
+	return ottl.StandardGetSetter[K]{
+		Getter: func(ctx K) interface{} {
 			return ctx.GetInstrumentationScope().Name()
 		},
-		Setter: func(ctx ottl.TransformContext, val interface{}) {
+		Setter: func(ctx K, val interface{}) {
 			if str, ok := val.(string); ok {
 				ctx.GetInstrumentationScope().SetName(str)
 			}
@@ -93,12 +97,12 @@ func accessInstrumentationScopeName() ottl.StandardGetSetter {
 	}
 }
 
-func accessInstrumentationScopeVersion() ottl.StandardGetSetter {
-	return ottl.StandardGetSetter{
-		Getter: func(ctx ottl.TransformContext) interface{} {
+func accessInstrumentationScopeVersion[K InstrumentationScopeContext]() ottl.StandardGetSetter[K] {
+	return ottl.StandardGetSetter[K]{
+		Getter: func(ctx K) interface{} {
 			return ctx.GetInstrumentationScope().Version()
 		},
-		Setter: func(ctx ottl.TransformContext, val interface{}) {
+		Setter: func(ctx K, val interface{}) {
 			if str, ok := val.(string); ok {
 				ctx.GetInstrumentationScope().SetVersion(str)
 			}