--- a/processor/transformprocessor/internal/logs/processor.go
+++ b/processor/transformprocessor/internal/logs/processor.go
@@ -26,7 +26,7 @@ import (
 )
 
 type Processor struct {
-	queries []ottl.Statement
+	queries []ottl.Statement[ottllogs.TransformContext]
 	logger  *zap.Logger
 }