… pkg/ottl/boolean_value_test.go
… pkg/ottl/compare_test.go
… pkg/ottl/contexts/internal/ottlcommon/resource_test.go
… pkg/ottl/contexts/internal/ottlcommon/scope_test.go
… pkg/ottl/contexts/ottldatapoints/datapoints.go
… pkg/ottl/contexts/ottltraces/traces.go
… pkg/ottl/expression.go
… pkg/ottl/expression_test.go
… pkg/ottl/functions.go
… pkg/ottl/functions_test.go
… pkg/ottl/ottlfuncs/func_concat_test.go
… pkg/ottl/ottlfuncs/func_delete_key_test.go
… pkg/ottl/ottlfuncs/func_delete_matching_keys.go
… pkg/ottl/ottlfuncs/func_delete_matching_keys_test.go
… pkg/ottl/ottlfuncs/func_int.go
… pkg/ottl/ottlfuncs/func_int_test.go
… pkg/ottl/ottlfuncs/func_is_match.go
… pkg/ottl/ottlfuncs/func_is_match_test.go
… pkg/ottl/ottlfuncs/func_keep_keys.go
… pkg/ottl/ottlfuncs/func_keep_keys_test.go
… pkg/ottl/ottlfuncs/func_limit.go
… pkg/ottl/ottlfuncs/func_limit_test.go
… pkg/ottl/ottlfuncs/func_replace_all_matches.go
… pkg/ottl/ottlfuncs/func_replace_all_matches_test.go
… pkg/ottl/ottlfuncs/func_replace_all_patterns.go
… pkg/ottl/ottlfuncs/func_replace_all_patterns_test.go
… pkg/ottl/ottlfuncs/func_replace_match.go
… pkg/ottl/ottlfuncs/func_replace_match_test.go
… pkg/ottl/ottlfuncs/func_replace_pattern.go
… pkg/ottl/ottlfuncs/func_replace_pattern_test.go