This PR also improves number of allocations (by 2, if I am not wrong is because of the removing of the ottl.TransformContext interface use, which requires a heap allocation) and overall performance of the transform processor.

**Before:**

```shell
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/processor/transformprocessor/internal/traces
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkTwoSpans
BenchmarkTwoSpans/no_processing
BenchmarkTwoSpans/no_processing-16         	  783144	      1394 ns/op	    1528 B/op	      31 allocs/op
BenchmarkTwoSpans/set_attribute
BenchmarkTwoSpans/set_attribute-16         	  713299	      1691 ns/op	    1832 B/op	      35 allocs/op
BenchmarkTwoSpans/keep_keys_attribute
BenchmarkTwoSpans/keep_keys_attribute-16   	  789517	      1626 ns/op	    1560 B/op	      33 allocs/op
BenchmarkTwoSpans/no_match
BenchmarkTwoSpans/no_match-16              	  799210	      1520 ns/op	    1560 B/op	      33 allocs/op
BenchmarkTwoSpans/inner_field
BenchmarkTwoSpans/inner_field-16           	  737241	      1566 ns/op	    1560 B/op	      33 allocs/op
BenchmarkTwoSpans/inner_field_both_spans
BenchmarkTwoSpans/inner_field_both_spans-16         	  763725	      1622 ns/op	    1592 B/op	      35 allocs/op
```

**After:**

```shell
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/processor/transformprocessor/internal/traces
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkTwoSpans
BenchmarkTwoSpans/no_processing
BenchmarkTwoSpans/no_processing-16         	  868864	      1296 ns/op	    1480 B/op	      29 allocs/op
BenchmarkTwoSpans/set_attribute
BenchmarkTwoSpans/set_attribute-16         	  738514	      1609 ns/op	    1784 B/op	      33 allocs/op
BenchmarkTwoSpans/keep_keys_attribute
BenchmarkTwoSpans/keep_keys_attribute-16   	  800833	      1450 ns/op	    1512 B/op	      31 allocs/op
BenchmarkTwoSpans/no_match
BenchmarkTwoSpans/no_match-16              	  835160	      1394 ns/op	    1512 B/op	      31 allocs/op
BenchmarkTwoSpans/inner_field
BenchmarkTwoSpans/inner_field-16           	  796388	      1445 ns/op	    1512 B/op	      31 allocs/op
BenchmarkTwoSpans/inner_field_both_spans
BenchmarkTwoSpans/inner_field_both_spans-16         	  798840	      1539 ns/op	    1544 B/op	      33 allocs/op
```